+++ src/network_event/server_event.rs
@@ -24,27 +24,32 @@ pub trait ServerEventAppExt {
         &mut self,
     ) -> &mut Self;
 
-    /// Same as [`Self::add_server_event`], but uses the specified receiving system.
-    fn add_server_event_with<T: Event + Serialize + DeserializeOwned + Debug, Marker>(
+    /// Same as [`Self::add_server_event`], but uses specified sending and receiving systems.
+    fn add_server_event_with<T: Event + Debug, Marker1, Marker2>(
         &mut self,
-        receiving_system: impl IntoSystemConfig<Marker>,
+        sending_system: impl IntoSystemConfig<Marker1>,
+        receiving_system: impl IntoSystemConfig<Marker2>,
     ) -> &mut Self;
 }
 
 impl ServerEventAppExt for App {
     fn add_server_event<T: Event + Serialize + DeserializeOwned + Debug>(&mut self) -> &mut Self {
-        self.add_server_event_with::<T, _>(receiving_system::<T>)
+        self.add_server_event_with::<T, _, _>(sending_system::<T>, receiving_system::<T>)
     }
 
     fn add_mapped_server_event<T: Event + Serialize + DeserializeOwned + Debug + MapEntities>(
         &mut self,
     ) -> &mut Self {
-        self.add_server_event_with::<T, _>(receiving_and_mapping_system::<T>)
+        self.add_server_event_with::<T, _, _>(
+            sending_system::<T>,
+            receiving_and_mapping_system::<T>,
+        )
     }
 
-    fn add_server_event_with<T: Event + Serialize + DeserializeOwned + Debug, Marker>(
+    fn add_server_event_with<T: Event + Debug, Marker1, Marker2>(
         &mut self,
-        receiving_system: impl IntoSystemConfig<Marker>,
+        sending_system: impl IntoSystemConfig<Marker1>,
+        receiving_system: impl IntoSystemConfig<Marker2>,
     ) -> &mut Self {
         let channel_id = self
             .world
@@ -57,11 +62,10 @@ impl ServerEventAppExt for App {
             .add_system(receiving_system.in_set(OnUpdate(ClientState::Connected)))
             .add_systems(
                 (
-                    sending_system::<T>.in_set(OnUpdate(ServerState::Hosting)),
-                    local_resending_system::<T>
-                        .after(sending_system::<T>)
-                        .in_set(ServerSet::Authority),
+                    sending_system.in_set(OnUpdate(ServerState::Hosting)),
+                    local_resending_system::<T>.in_set(ServerSet::Authority),
                 )
+                    .chain()
                     .in_set(ServerSet::Tick),
             );